Added test for replicationTaskExecutor
